requirements.txt
# Pickr AI Email Sales Rep v2 - Dependencies

# Core
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1

# Database (PostgreSQL)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Redis (Job Queue)
redis==5.0.1

# Web scraping & enrichment
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0

# AI / LLM
anthropic==0.43.0

# Templating (for dashboard)
jinja2==3.1.3

# Utilities
python-multipart==0.0.9
rich==13.7.0
